28-aboc-upgraded/requirements.txt

langchain==0.1.16
langgraph==0.0.40
chromadb==0.4.24
openai==1.14.3
tiktoken==0.6.0
python-dotenv==1.0.1

fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.33.0
requests==2.31.0